+++ src/Bootstrap/ReplaceTemplateTags.php
@@ -24,6 +24,7 @@ class ReplaceTemplateTags
             $exp->registerTag('channel.entries', new \Expressionengine\Coilpack\View\Tags\Channel\Entries);
             $exp->registerTag('comment.entries', new \Expressionengine\Coilpack\View\Tags\Comment\Entries);
             $exp->registerTag('structure.entries', new \Expressionengine\Coilpack\View\Tags\Structure\Entries);
+            $exp->registerTag('structure.nav', new \Expressionengine\Coilpack\View\Tags\Structure\Nav);
             $exp->registerTag('email.contact_form', new \Expressionengine\Coilpack\View\Tags\Email\ContactForm);
         });